Rename to mongoid-enum-dazzl
