+++ lib/mongoid/enum/dazzl/configuration.rb
@@ -1,0 +1,23 @@
+# frozen_string_literal: true
+
+module Mongoid
+  module Enum
+    module Dazzl
+      class Configuration
+        attr_accessor :field_name_prefix
+
+        def initialize
+          self.field_name_prefix = '_'
+        end
+      end
+
+      def self.configuration
+        @configuration ||= Configuration.new
+      end
+
+      def self.configure
+        yield(configuration) if block_given?
+      end
+    end
+  end
+end
