--- lib/mongoid7/enum/configuration.rb
@@ -1,21 +1,0 @@
-# frozen_string_literal: true
-
-module Mongoid7
-  module Enum
-    class Configuration
-      attr_accessor :field_name_prefix
-
-      def initialize
-        self.field_name_prefix = '_'
-      end
-    end
-
-    def self.configuration
-      @configuration ||= Configuration.new
-    end
-
-    def self.configure
-      yield(configuration) if block_given?
-    end
-  end
-end
