+++ spec/mongoid/enum/dazzl_spec.rb
@@ -1,18 +1,18 @@
 # frozen_string_literal: true
 
 require 'spec_helper'
-require 'mongoid7/enum/configuration'
+require 'mongoid/enum/dazzl/configuration'
 
 class User
   include Mongoid::Document
-  include Mongoid7::Enum
+  include Mongoid::Enum::Dazzl
 
   enum :status, %i[awaiting_approval approved banned]
   enum :roles, %i[author editor admin], multiple: true, default: [], required: false
 end
 
 # rubocop:disable RSpec/LeakyConstantDeclaration, Lint/ConstantDefinitionInBlock
-describe Mongoid7::Enum do
+describe Mongoid::Enum::Dazzl do
   let(:klass) { User }
   let(:instance) { User.new }
   let(:alias_name) { :status }
@@ -34,7 +34,7 @@ describe Mongoid7::Enum do
 
       UserWithoutPrefix = Class.new do
         include Mongoid::Document
-        include Mongoid7::Enum
+        include Mongoid::Enum::Dazzl
 
         enum :status, %i[awaiting_approval approved banned]
       end
